Proteksi Modul using middleware

+++ resources/views/siswa/index.blade.php
@@ -26,17 +26,19 @@
                             <td>{{ $siswa->jenis_kelamin }}</td>
                             <td>{{ !empty($siswa->telepon->nomor_telepon) ? $siswa->telepon->nomor_telepon : '-' }}</td>
                             <td>
-                                <div class="box-button">
-                                    {{ link_to('siswa/'.$siswa->id,'Detail',['class' => 'btn btn-success btn-sm'])}}
-                                </div>
-                                <div class="box-button">
-                                    {{ link_to('siswa/'.$siswa->id.'/edit','Edit',['class' => 'btn btn-warning btn-sm'])}}
-                                </div>
-                                <div class="box-button">
-                                    {!! Form::model($siswa,['method' => 'DELETE','action' => ['SiswaController@destroy',$siswa->id]]) !!}
-                                        {!! Form::submit('Delete', ['class' => 'btn btn-danger btn-sm']) !!}
-                                    {!! Form::close() !!}
-                                </div>
+                                @if (Auth::check())
+                                    <div class="box-button">
+                                        {{ link_to('siswa/'.$siswa->id,'Detail',['class' => 'btn btn-success btn-sm'])}}
+                                    </div>
+                                    <div class="box-button">
+                                        {{ link_to('siswa/'.$siswa->id.'/edit','Edit',['class' => 'btn btn-warning btn-sm'])}}
+                                    </div>
+                                    <div class="box-button">
+                                        {!! Form::model($siswa,['method' => 'DELETE','action' => ['SiswaController@destroy',$siswa->id]]) !!}
+                                            {!! Form::submit('Delete', ['class' => 'btn btn-danger btn-sm']) !!}
+                                        {!! Form::close() !!}
+                                    </div>
+                                @endif
                             </td>
                         </tr>
                     @endforeach
@@ -54,11 +56,13 @@
                 {{ $siswa_list->links() }}
             </div>
         </div>
-        <div class="botton-nav">
-            <div>
-                <a href="{{ route('siswa.create') }}" class="btn btn-primary">Tambah</a>
+        @if (Auth::check())
+            <div class="botton-nav">
+                <div>
+                    <a href="{{ route('siswa.create') }}" class="btn btn-primary">Tambah</a>
+                </div>
             </div>
-        </div>
+        @endif
     </div>
 @stop
 @section('footer')
